Implemented More Controllers

+++ src/main/java/iyteyazilim/projects/digitalcard/repository/IAnnouncementRepository.java
@@ -3,5 +3,5 @@ package iyteyazilim.projects.digitalcard.repository;
 import iyteyazilim.projects.digitalcard.entity.Announcement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IAnnouncementRepository extends JpaRepository<Announcement, Integer> {
+public interface IAnnouncementRepository extends JpaRepository<Announcement, Long> {
 }
