+++ src/main/java/iyteyazilim/projects/digitalcard/repository/ICommentRepository.java
@@ -3,5 +3,5 @@ package iyteyazilim.projects.digitalcard.repository;
 import iyteyazilim.projects.digitalcard.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICommentRepository extends JpaRepository<Comment, Integer> {
+public interface ICommentRepository extends JpaRepository<Comment, Long> {
 }
