+++ src/main/java/iyteyazilim/projects/digitalcard/repository/IEventRepository.java
@@ -3,5 +3,5 @@ package iyteyazilim.projects.digitalcard.repository;
 import iyteyazilim.projects.digitalcard.entity.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IEventRepository extends JpaRepository<Event, Integer> {
+public interface IEventRepository extends JpaRepository<Event, Long> {
 }
